test: refactor to use test-utils
