+++ test/e2e/fixtures/pool-worker-exit/1-first.test.ts
@@ -1,0 +1,6 @@
+import { expect, test } from 'vitest'
+import { covered } from './src'
+
+test('first test exercises src so it should appear in coverage', () => {
+  expect(covered()).toBe(42)
+})
