+++ test/e2e/fixtures/pool-worker-exit/2-crash.test.ts
