+++ test/e2e/fixtures/pool-worker-exit/3-second.test.ts
