--- test/e2e/fixtures/pool-worker-exit/first.test.js
@@ -1,6 +1,0 @@
-import { expect, test } from 'vitest'
-import { covered } from './src.js'
-
-test('first test exercises src.js so it should appear in coverage', () => {
-  expect(covered()).toBe(42)
-})
