--- test/e2e/fixtures/pool-worker-exit/src.js
@@ -1,3 +1,0 @@
-export function covered() {
-  return 42
-}
