+++ test/e2e/fixtures/pool-worker-exit/src.ts
@@ -1,0 +1,7 @@
+export function covered() {
+  return 42
+}
+
+export function uncovered() {
+  return 42
+}
